modify itoa

+++ ycdatabase_extension/include/ycdatabase.h
@@ -143,7 +143,7 @@ extern char* ltrim(char *str);
 extern char* trim(char *str);
 extern char* ltrim_str(char *str, char *remove);
 extern char* rtrim_str(char *str, char *remove);
-extern char* yc_itoa(int num,char* str);
+extern char* yc_itoa(long num,char* str);
 extern char* yc_escape(char* sql);
 extern char* strreplace(char* original, char const * const pattern, char const * const replacement);
 extern void yc_serialize(zval *return_value, zval *var TSRMLS_DC);
